+++ knf/knf_test.go
@@ -16,6 +16,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/essentialkaos/ek/v13/errors"
+
 	check "github.com/essentialkaos/check"
 )
 
@@ -231,7 +233,7 @@ func (s *KNFSuite) TestErrors(c *check.C) {
 	c.Assert(HasProp("test:test"), check.Equals, false)
 	c.Assert(Sections(), check.HasLen, 0)
 	c.Assert(Props("test"), check.HasLen, 0)
-	c.Assert(Validate([]*Validator{}), check.DeepEquals, []error{ErrNilConfig})
+	c.Assert(Validate(Validators{}), check.DeepEquals, errors.Errors{ErrNilConfig})
 	c.Assert(Alias("test:test", "test:test"), check.NotNil)
 	c.Assert(global.Merge(nil), check.NotNil)
 
@@ -253,7 +255,7 @@ func (s *KNFSuite) TestErrors(c *check.C) {
 	c.Assert(config.HasProp("test:test"), check.Equals, false)
 	c.Assert(config.Sections(), check.HasLen, 0)
 	c.Assert(config.Props("test"), check.HasLen, 0)
-	c.Assert(config.Validate([]*Validator{}), check.HasLen, 0)
+	c.Assert(config.Validate(Validators{}), check.HasLen, 0)
 	c.Assert(config.Merge(nil), check.NotNil)
 
 	c.Assert(config.Alias("", ""), check.NotNil)
@@ -620,10 +622,10 @@ func (s *KNFSuite) TestNil(c *check.C) {
 	c.Assert(err, check.NotNil)
 	c.Assert(err, check.DeepEquals, ErrNilConfig)
 
-	errs := nilConf.Validate([]*Validator{})
+	errs := nilConf.Validate(Validators{})
 
 	c.Assert(errs, check.Not(check.HasLen), 0)
-	c.Assert(errs, check.DeepEquals, []error{ErrNilConfig})
+	c.Assert(errs, check.DeepEquals, errors.Errors{ErrNilConfig})
 }
 
 func (s *KNFSuite) TestDefault(c *check.C) {
@@ -701,11 +703,21 @@ func (s *KNFSuite) TestSimpleValidator(c *check.C) {
 		return nil
 	}
 
-	errs := Validate([]*Validator{
+	validators := Validators{
+		{"string:test2", simpleValidator, nil},
+	}
+
+	validators = validators.AddIf(false, Validators{
 		{"string:test2", simpleValidator, nil},
 	})
 
-	c.Assert(errs, check.HasLen, 1)
+	validators = validators.AddIf(true, Validators{
+		{"string:test2", simpleValidator, nil},
+	})
+
+	errs := Validate(validators)
+
+	c.Assert(errs, check.HasLen, 2)
 }
 
 func (s *KNFSuite) TestKNFParserExceptions(c *check.C) {